Moved log data types and classes to a logs directory.
